+++ packages/widgetbook/test/src/addons/device_frame_addon_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:widgetbook/widgetbook.dart';
+import 'package:widgetbook/src/widgetbook_theme.dart';
+import 'package:widgetbook/widgetbook.dart' hide WidgetbookTheme;
 
 import '../../helper/helper.dart';
 
@@ -162,6 +163,36 @@ void main() {
             ),
             findsOneWidget,
           );
+        },
+      );
+
+      testWidgets(
+        'given a use-case, '
+        'then the background is set to [scaffoldBackgroundColor]',
+        (tester) async {
+          final device = devices.last;
+          const color = Color(0xff123456);
+
+          await tester.pumpWidgetWithBuilder(
+            (context) => WidgetbookTheme(
+              data: Theme.of(context).copyWith(
+                scaffoldBackgroundColor: color,
+              ),
+              child: Builder(
+                builder: (context) {
+                  return addon.buildUseCase(
+                    context,
+                    const SizedBox(),
+                    DeviceFrameSetting(
+                      device: device,
+                    ),
+                  );
+                },
+              ),
+            ),
+          );
+
+          expect(find.byType(ColoredBox), paints..rect(color: color));
         },
       );
     },