+++ packages/widgetbook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FIX**: Unify background color between the workbench and `DeviceFrameAddon`. ([#1461](https://github.com/widgetbook/widgetbook/pull/1461) - by [@ValentinVignal](https://github.com/ValentinVignal))
+
 ## 3.14.2
 
 - **FIX**: Tighten `device_frame_plus` version constraints; to prevent the [breaking changes](https://github.com/StorybookToolkit/device_frame_plus/issues/5#issuecomment-2907648670) introduced in v1.4.0. ([#1475](https://github.com/widgetbook/widgetbook/pull/1475) - by [@Peetee06](https://github.com/Peetee06))